+++ logstash/plugins/logstash-filter-urlgrok/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.4
+ - By default the tag_prefix has been set to nil has to be defined in the filter config if you would like to debug events
+## 1.0.3
+ - Fixed bug that add blank category if the segemnet was null
 ## 1.0.2
  - Added functionality to define multiple tags for the category tags
 ## 1.0.1